Db latest (#89)
* refactor: update gems fields to Long for consistency

* refactor: deleted allowedUsers unused entity and repository

* refactor: changed Long datatype to UUID for every entities and related dtos-service-controller-tests

* feat:add createdAt updatedAt timestamp in Program entity

* feat:add createdAt updatedAt timestamp in User entity

* feat:add createdAt updatedAt timestamp in UserWallet entity

* refactor: changed datatype of userid to long

* refactor: refactored test as per user id datatype Long

* pom change

* refactor: add userid auto increment

* modify: changed path for logs folder

* modify: removed ddl-auto from yml

+++ questevent-backend/src/main/java/com/questevent/entity/ProgramWallet.java
@@ -38,7 +38,7 @@ public class ProgramWallet {
     private User user;
 
     @Column(name = "gems", nullable = false)
-    private int gems;
+    private Long gems;
 
     @Enumerated(EnumType.STRING)
     private ProgramStatus status;
